// src/components/Autenticazione/types/form-values.type.ts
export interface IFormValues {
  username: string;
  phone: string;
  email: string;
  password: string;
  /*  successful: false,
      message: string, */
  name: string;
  surname: string;
  fiscalCode: string;
  genre: string;
  birthday: string;
  address: string;
  houseNumber: string;
  province: string;
  municipality: string;
  cap: string;
  issueEntity: string;
  documentType: string;
  issueDate: string;
  documentNumber: string;
  issuePlace: string;
  terminiCondizioni: boolean;
  statutoPrivacy: boolean;
  faceFile: string;
  role: number;
  //Tipo file da caricare, da vedere di quale tipo si tratta
  //documentFile: string;
}

/* export interface IResponse{
    return_value
} */